Patch cel-cpp, fixing cilium-proxy build
We're hitting this cel-cpp issue while buildig cilium-proxy:

* https://github.com/envoyproxy/envoy/issues/34368

Bumping cel-cpp to 0.10.0 has been suggested, however that would
require bumping other dependencies as well (e.g. com_google_absl).

Instead, we'll apply the fix that moves the offending cel-cpp
header definitions:

https://github.com/google/cel-cpp/commit/18cf5685e93fb8c07c807d913fabe9d1b54db87f

Applying this patch is not straight forward because the build
process is based on bazel and heavily nested: the rockcraft
script will patch the cilium-proxy build scripts, which in turn
need to patch the envoy proxy build scripts to patch the cel-cpp
dependency.